Move InstancesRepositoryTest
