+++ androidshared/.gitignore
