+++ androidshared/src/main/AndroidManifest.xml
