+++ collect_app/build.gradle
@@ -334,7 +334,7 @@ dependencies {
     }
 
     implementation project(':shared')
-    implementation project(':android-shared')
+    implementation project(':androidshared')
     implementation project(':nbistubs')
     implementation project(':strings')
     implementation project(':material')
