+++ collect_app/src/test/java/org/odk/collect/android/database/DatabaseInstancesRepositoryTest.java
@@ -19,7 +19,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.odk.collect.forms.instances.InstancesRepository;
-import org.odk.collect.android.instances.InstancesRepositoryTest;
+import org.odk.collect.forms.testsupport.InstancesRepositoryTest;
 import org.odk.collect.android.storage.StorageInitializer;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
