+++ collect_app/src/test/java/org/odk/collect/android/formmanagement/FormDeleterTest.java
@@ -7,7 +7,7 @@ import org.odk.collect.android.itemsets.FastExternalItemsetsRepository;
 import org.odk.collect.forms.testsupport.FormUtils;
 import org.odk.collect.android.support.InMemFormsRepository;
 import org.odk.collect.android.support.InMemInstancesRepository;
-import org.odk.collect.testshared.TempFiles;
+import org.odk.collect.shared.TempFiles;
 
 import java.io.File;
 import java.util.List;
@@ -16,7 +16,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
-import static org.odk.collect.android.support.InstanceUtils.buildInstance;
+import static org.odk.collect.forms.testsupport.InstanceUtils.buildInstance;
 
 
 public class FormDeleterTest {
