+++ collect_app/src/test/java/org/odk/collect/android/instancemanagement/InstanceDeleterTest.java
@@ -7,13 +7,13 @@ import org.odk.collect.forms.instances.Instance;
 import org.odk.collect.forms.testsupport.FormUtils;
 import org.odk.collect.android.support.InMemFormsRepository;
 import org.odk.collect.android.support.InMemInstancesRepository;
-import org.odk.collect.testshared.TempFiles;
+import org.odk.collect.shared.TempFiles;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
-import static org.odk.collect.android.support.InstanceUtils.buildInstance;
+import static org.odk.collect.forms.testsupport.InstanceUtils.buildInstance;
 
 public class InstanceDeleterTest {
