+++ collect_app/src/test/java/org/odk/collect/android/provider/InstanceProviderTest.java
@@ -14,7 +14,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.storage.StorageInitializer;
-import org.odk.collect.android.support.InstanceUtils;
+import org.odk.collect.forms.testsupport.InstanceUtils;
 import org.robolectric.shadows.ShadowEnvironment;
 
 import java.io.File;
