+++ collect_app/src/test/java/org/odk/collect/android/support/InMemFormsRepository.java
@@ -5,7 +5,7 @@ import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.FormsRepository;
 import org.odk.collect.shared.Md5;
-import org.odk.collect.testshared.TempFiles;
+import org.odk.collect.shared.TempFiles;
 
 import java.io.File;
 import java.io.IOException;
